Only component renamed :)
